+++ test/field.campaign-code.spec.js
@@ -35,7 +35,7 @@ describe('field.campaign-code', () => {
 
     it('translate default', async () => {
       let r = await f.convert('campaigncode', {code: 'test', _parent: 'test'}, logger);
-      assert.equal(r.typeId, 62, 'found default');
+      assert.equal(r.typeId, 0, 'found default');
       assert.equal(r._parent, 'test', 'did store test')
     });
     it('translate value', async () => {